Add files via upload
The r shiny app that shows the temperature statistics of stations in Nepal from the year of 2013 to 2017. The dashboard has two tables. 
The first table shows the maximum, minimum and average temperatures of all the stations in Nepal from 2013 to 2017. 
The second table shows the Latitude and Longitude of the stations.
The app also has a map of Nepal on the bottom where markers are pinned to show the location of the stations. When you click the markers, it shows the name of the station. 
The map also shows a single location when you select a station from the second table. But only one station can be selected.